+++ app-backend/datamodel/src/main/scala/UserIntercomConversation.scala
@@ -1,0 +1,6 @@
+package com.rasterfoundry.datamodel
+
+final case class UserIntercomConversation(
+    userId: String,
+    conversationId: String
+)