+++ app-backend/notification/src/main/scala/com/rasterfoundry/notification/intercom/Model.scala
@@ -14,49 +14,77 @@ object Model {
       Decoder.decodeString.map(ExternalId.apply _)
   }
 
-  @newtype case class Notification(underlying: String)
-  @newtype case class UserId(underlying: String)
-  object UserId {
-    implicit val encUserId: Encoder[UserId] =
+  @newtype case class ConversationId(underlying: String)
+  object ConversationId {
+    implicit val encConversationId: Encoder[ConversationId] =
       Encoder.encodeString.contramap(_.underlying)
-    implicit val decUserId: Decoder[UserId] =
-      Decoder.decodeString.map(UserId.apply _)
+    implicit val decConversationId: Decoder[ConversationId] =
+      Decoder.decodeString.map(ConversationId.apply _)
   }
+
+  @newtype case class Notification(underlying: String)
   @newtype case class Message(underlying: String)
   @newtype case class IntercomToken(underlying: String)
 
-  case class FromObject(adminId: UserId)
-  object FromObject {
-    implicit val encFromObject: Encoder[FromObject] =
-      Encoder.forProduct2("type", "id")(
-        fromObject => ("admin", fromObject.adminId)
-      )
+  @newtype case class AdminId(underlying: String)
+  object AdminId {
+    implicit val encAdminId: Encoder[AdminId] =
+      Encoder.encodeString.contramap(_.underlying)
+    implicit val decAdminId: Decoder[AdminId] =
+      Decoder.decodeString.map(AdminId.apply _)
   }
 
-  case class ToObject(externalId: ExternalId)
-  object ToObject {
-    implicit val encToObject: Encoder[ToObject] = Encoder.forProduct2(
+  case class AdminObject(adminId: AdminId)
+  object AdminObject {
+    implicit val encAdminObject: Encoder[AdminObject] =
+      Encoder.forProduct2("type", "id")(adminObject =>
+        ("admin", adminObject.adminId))
+  }
+
+  case class UserObject(externalId: ExternalId)
+  object UserObject {
+    implicit val encUserObject: Encoder[UserObject] = Encoder.forProduct2(
       "type",
       "user_id"
-    )(toObject => ("user", toObject.externalId))
+    )(userObject => ("user", userObject.externalId))
   }
 
-  case class MessagePost(adminId: UserId, userId: ExternalId, msg: Message)
-  object MessagePost {
-    implicit val encMessagePost: Encoder[MessagePost] = Encoder.forProduct4(
-      "from",
-      "to",
-      "body",
-      "message_type"
-    )(
-      messagePost =>
-        (
-          FromObject(messagePost.adminId),
-          ToObject(messagePost.userId),
-          messagePost.msg.underlying,
-          "inapp"
-      )
-    )
+  case class ConversationCreate(userId: ExternalId, message: Message)
+  object ConversationCreate {
+    implicit val encConversationCreate: Encoder[ConversationCreate] =
+      Encoder.forProduct2(
+        "from",
+        "body"
+      )(
+        conversationCreate =>
+          (
+            UserObject(conversationCreate.userId),
+            conversationCreate.message.underlying
+        ))
   }
 
+  case class Conversation(conversationId: ConversationId)
+  object Conversation {
+    implicit val decConversation: Decoder[Conversation] =
+      Decoder.forProduct1("conversation_id")((conversationId: String) =>
+        Conversation(ConversationId(conversationId)))
+  }
+
+  case class ConversationReply(adminId: AdminId, message: Message)
+  object ConversationReply {
+    implicit val encConversationReply: Encoder[ConversationReply] =
+      Encoder.forProduct4(
+        "message_type",
+        "type",
+        "admin_id",
+        "body"
+      )(
+        conversationReply =>
+          (
+            "comment",
+            "admin",
+            conversationReply.adminId.underlying,
+            conversationReply.message.underlying
+        ))
+  }
 }