+++ app-tasks/rf/src/rf/uploads/geotiff/__init__.py
@@ -1,4 +1,3 @@
-from .create_thumbnails import create_thumbnails
 from .create_scenes import create_geotiff_scene
 from .factories import GeoTiffS3SceneFactory
 from .create_images import create_geotiff_image